src/requirements.txt
fastapi
uvicorn
requests
pydantic
openai>=1.0.0
python-dotenv
